We corrected the program; we confused its purpose. Correction: model, repository, and BookService

+++ src/main/java/com/alura/literalura/repository/AuthorRepository.java
@@ -1,9 +1,9 @@
 package com.alura.literalura.repository;
 
-import com.alura.literalura.model.Author;
 import org.springframework.data.jpa.repository.JpaRepository;
+import com.alura.literalura.model.Author;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AuthorRepository extends JpaRepository<Author, Long> {
+public interface AuthorRepository extends JpaRepository<Author, Integer> {
 }
